Edits to 1.1 From candidate to declared paper
In general, I think this section can be condensed, as it goes too deep into the nitty-gritty of the candidate to declared paper transition. That should probably be included on the platform and/or streamlined into the form. Not to downplay the importance of selecting a paper, the bulk of this step should be about declaring the outputs that will be analyzed in the remainder of the exercise. 

Here's a summary of my edits and a few comments:
- Removed description of candidate vs declared paper in the first paragraph (this is already in the intro paragraphs in the edited version)
- The decision tree does not show up. Is this something that is being developed still? Nonetheless, I'm not sure we need a separate section for it (it breaks off the section 1.1 From candidate to declared paper into a sub-section, but we don't do that for the other 3 steps of the process of verifying whether a reproduction package exists).
- 1.1.3 Decision Tree to Move From Candidate to Declared Paper is very difficult to follow and can be condensed (I'll do that in my next round of edits).