Adding a note under 3.1 Describe the inputs
Maybe adding a note could be helpful in section "3.1 Describe the inputs" that explains that not all replication packages include analysis data files, some use the raw data as inputs to produce outputs directly. Not all replication packages include raw data, sometimes there may be access restrictions, etc. such that raw data is not allowed to be redistributed (which is explained later in the chapter).