Sources for identifying a proper citation
It might be helpful to explain what a proper citation is, some papers have in text citations, others only explain the data but do not cite in the references section, etc. so a framework for citations may be helpful to determine whether or not the source is properly cited. This link has some good examples: https://www.aeaweb.org/journals/policies/sample-references