Downloading and uploading the CSV file
Would be helpfult to add a note saying replicators can download the table as CSV file, complete recording all the information there and upload the final version into the form. Sometimes there are upwards of 15 datasets and it seems to be easier to do it in excel than in the survey for such large volume.